+++ src/app/dishes/[id]/dish-image.tsx
@@ -1,9 +1,9 @@
 'use client';
 
-import styles from '@/styles/app/result/page.module.scss';
+import styles from '@/styles/app/dishes/[id]/page.module.scss';
 import { useHasImage } from '@/src/hooks/use-has-image';
 import { Loader } from '@/src/components/common/loader';
-import { getImageUrlForYummyDishes } from '@/src/helpers/result.helper';
+import { getImageUrlForYummyDishes } from '@/src/helpers/dish-by-id.helper';
 import { DishProvider } from '@/src/types/api.types';
 
 interface DishImageProps {