+++ src/app/recommendations/recommendations-navigator.tsx
@@ -1,6 +1,6 @@
 'use client';
 
-import styles from '@/styles/app/dish-proposal/page.module.scss';
+import styles from '@/styles/app/recommendations/page.module.scss';
 import { Button } from '@/src/components/common/button';
 
 interface DishProposalNavigatorProps {
@@ -8,9 +8,9 @@ interface DishProposalNavigatorProps {
     onNext: () => void;
 }
 
-export function DishProposalNavigator({ onPrevious, onNext }: DishProposalNavigatorProps) {
+export function RecommendationsNavigator({ onPrevious, onNext }: DishProposalNavigatorProps) {
     return (
-        <div className={styles['dish-proposal-container__navigator']}>
+        <div className={styles['recommendations-container__navigator']}>
             <Button label="Poprzednia" icon="previous" onClick={onPrevious} width={200} />
             <Button label="Następna" icon="next" onClick={onNext} width={200} />
         </div>