+++ src/app/users/registration/registration-form.tsx
@@ -6,12 +6,15 @@ import styles from '@/styles/app/users/registration/registration-form.module.scs
 import { InputString } from '@/src/components/common/form/input-string';
 import { Button } from '@/src/components/common/button';
 import { InputPassword } from '@/src/components/common/form/input-password';
-import { BackLink } from '@/src/components/common/back-link';
-import { useEffect, useMemo, useState } from 'react';
+import { useEffect, useState } from 'react';
 import { changeUserPassword, createUserAccount } from '@/src/api/api';
 import { toastError, toastSuccess } from '@/src/utils/toast.utils';
 import { Loader } from '@/src/components/common/loader';
-import { useRouter, useSearchParams } from 'next/navigation';
+import { useRouter } from 'next/navigation';
+
+interface RegistrationFormProps {
+    isResetting?: boolean;
+}
 
 const defaultValues: UserData = {
     email: '',
@@ -20,10 +23,8 @@ const defaultValues: UserData = {
     repeatedPassword: ''
 };
 
-export function RegistrationForm() {
+export function RegistrationForm({ isResetting }: RegistrationFormProps) {
     const router = useRouter();
-    const searchParams = useSearchParams();
-    const isResetting = useMemo(() => Boolean(searchParams.get('action') === 'reset'), [searchParams.get('action')]);
     const { handleSubmit, control, formState: { errors }, watch, reset, setValue, clearErrors } = useForm<UserData>({ defaultValues, mode: 'onChange' });
     const [isRegistering, setIsRegistering] = useState<boolean>(false);
     const [wasCreated, setWasCreated] = useState<boolean>(false);
@@ -88,10 +89,6 @@ export function RegistrationForm() {
     return (
         <form onSubmit={handleSubmit(isResetting ? onReset : onSubmit)} className={styles['registration-form']}>
             {isRegistering && <Loader isAbsolute={true} />}
-            {isResetting
-                ? <BackLink link="/" label={'Powrót do strony głównej'} isAttached={true} />
-                : <BackLink link="/users/login" label={'Powrót do logowania'} isAttached={true} />
-            }
             <h2>{isResetting ? 'Zmień hasło' : 'Stwórz nowe konto'}</h2>
             {!isResetting && (
                 <>